requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
boto3==1.28.62
neo4j==5.14.1
python-dotenv==1.0.0 